:tada:  重要功能, 样式, 数据, 结构基本完成(不过代码还没有优化 :poop:) :tada:	modified:   public/favicon.ico :tada:	modified:   public/index.html :tada:	new file:   public/logo192.png :tada:	new file:   public/logo512.png :tada:	new file:   public/manifest.json :tada:	modified:   src/App.css :tada:	modified:   src/App.js :tada:	modified:   src/App.scss :tada:	modified:   src/api/api.js :tada:	new file:   src/asserts/imgs/BackgroundImages/autumn.jpg :tada:	new file:   src/asserts/imgs/BackgroundImages/fancy.jpg :tada:	new file:   src/asserts/imgs/BackgroundImages/flower-field.jpg :tada:	new file:   src/asserts/imgs/BackgroundImages/hilltop.jpg :tada:	new file:   src/asserts/imgs/BackgroundImages/mountain.jpg :tada:	new file:   src/asserts/imgs/BackgroundImages/snow.jpg :tada:	new file:   src/asserts/imgs/BackgroundImages/sunset-glow.jpg :tada:	new file:   src/asserts/imgs/BackgroundImages/wriggle.jpg :tada:	modified:   src/asserts/imgs/SideBar/search_light.png :tada:	modified:   src/asserts/style/default.css :tada:	modified:   src/asserts/style/default.scss :tada:	new file:   src/asserts/style/theme.css :tada:	new file:   src/asserts/style/theme.scss :tada:	new file:   src/asserts/style/theme_option.css :tada:	new file:   src/asserts/style/theme_option.scss :tada:	modified:   src/components/ContextMenu/index.css :tada:	modified:   src/components/EmptyStatus/index.css :tada:	modified:   src/components/MainBody/index.css :tada:	modified:   src/components/MainBody/index.scss :tada:	modified:   src/components/SgCategoryItem/index.css :tada:	modified:   src/components/SgCategoryItem/index.jsx :tada:	modified:   src/components/SgPageTitle/index.css :tada:	modified:   src/components/SgRadioItem/index.css :tada:	modified:   src/components/SgSwitchItem/index.css :tada:	modified:   src/components/SgTodoItem/index.css :tada:	modified:   src/components/SideBar/index.css :tada:	modified:   src/components/SideBar/index.jsx :tada:	modified:   src/components/SideBar/index.scss :tada:	modified:   src/components/Switch/index.css :tada:	modified:   src/configs/config.js :tada:	modified:   src/index.css :tada:	modified:   src/index.js :tada:	modified:   src/index.scss :tada:	modified:   src/pages/Setting/index.css :tada:	modified:   src/pages/Setting/index.jsx :tada:	modified:   src/pages/Setting/index.scss :tada:	modified:   src/routes/index.js :tada:	modified:   src/utils/handleData.js

+++ src/App.js
@@ -14,8 +14,9 @@ import SideBar from './components/SideBar'
 import myRoutes from './routes'
 
 // 初始化本地储存
-import { initLocalData } from './api/api'
+import { changeToDoListBackgroundSetting, fetchConfirmBeforeDelete, initLocalData } from './api/api'
 import { alterTodoFinishedData, alterTodoImportantData, alterTodoTodayData, deleteData, fetchData } from './utils/handleData';
+import { getItem } from './utils/localStroge';
 initLocalData();
 
 
@@ -110,7 +111,17 @@ window.addEventListener('click', function (e) {
                 alterTodoFinishedData(tempMenu.route, tempMenu.id, !tempMenu.isFinished);
                 break;
             case 'delete':
-                deleteData(tempMenu.route, tempMenu.type, tempMenu.id);
+                // 调取设置
+                let setting = fetchConfirmBeforeDelete();
+
+                // 如果设置为真
+                if (setting.checked && setting.checked === true) {
+                    const flag = this.confirm('确认删除这条ToDo吗?');
+                    if (flag) deleteData(tempMenu.route, tempMenu.type, tempMenu.id);
+                } else {
+                    deleteData(tempMenu.route, tempMenu.type, tempMenu.id);
+                }
+
                 break;
 
             default:
@@ -134,6 +145,7 @@ export default function App() {
         })
 
         try {
+            changeToDoListBackgroundSetting(getItem('backgroundImg'));
             tempMenu.route = pathname.slice(1);
         } catch (error) { }
